+++ libs/react/src/context/coldComplianceQuestionnaireContext.ts
@@ -1,9 +1,10 @@
 import { createContext } from 'react';
-import { AIDetails, ComplianceSidebarPayload } from '@coldpbc/interfaces';
+import { Compliance, ComplianceSidebarPayload, AIDetails } from '@coldpbc/interfaces';
 import { SWRResponse } from 'swr';
 
 export interface ComplianceQuestionnaireContextType {
   name: string;
+  complianceDefinition: Compliance | undefined;
   scrollToQuestion: string | null;
   setScrollToQuestion: (question: string | null) => void;
   focusQuestion: {
@@ -21,6 +22,7 @@ export interface ComplianceQuestionnaireContextType {
 
 export const ColdComplianceQuestionnaireContext = createContext<ComplianceQuestionnaireContextType>({
   name: '',
+  complianceDefinition: undefined,
   scrollToQuestion: null,
   setScrollToQuestion: () => {},
   focusQuestion: null,